refactoring a small bit
