+++ src/main/java/com/joostmsoftware/startingrevamp/block/StartingRevampRocks.java
@@ -1,5 +1,6 @@
 package com.joostmsoftware.startingrevamp.block;
 
+import com.joostmsoftware.startingrevamp.block.custom.RockBlock;
 import com.joostmsoftware.startingrevamp.registry.StartingRevampBlockRegistry;
 import net.minecraft.block.Block;
