+++ src/main/java/com/joostmsoftware/startingrevamp/block/custom/TwigBlock.java
@@ -1,4 +1,4 @@
-package com.joostmsoftware.startingrevamp.block;
+package com.joostmsoftware.startingrevamp.block.custom;
 
 import net.fabricmc.fabric.api.object.builder.v1.block.FabricBlockSettings;
 import net.minecraft.block.Block;
